Refactor activities. Try unlocking screen
