+++ BankChain/app/src/androidTest/java/nl/tudelft/ewi/ds/bankchain/activities/MainActivityTest.java
@@ -1,12 +1,16 @@
-package nl.tudelft.ewi.ds.bankchain;
+package nl.tudelft.ewi.ds.bankchain.activities;
 
 import android.support.test.espresso.intent.rule.IntentsTestRule;
+import android.support.test.espresso.matcher.ViewMatchers;
 import android.support.test.runner.AndroidJUnit4;
+import android.view.WindowManager;
 
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import nl.tudelft.ewi.ds.bankchain.R;
 import nl.tudelft.ewi.ds.bankchain.activities.MainActivity;
 
 import static android.support.test.espresso.Espresso.onView;
@@ -32,10 +36,21 @@ public class MainActivityTest {
     public IntentsTestRule<MainActivity> mIntentsRule =
             new IntentsTestRule<>(MainActivity.class);
 
+    @Before
+    public void unlockScreen() {
+        final MainActivity activity = mIntentsRule.getActivity();
+
+        activity.runOnUiThread(() -> {
+            activity.getWindow().addFlags(WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON |
+                    WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED |
+                    WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+        });
+    }
+
     @Test
     public void verifyOpenRecentTransactionsActivity() {
         // Clicks a button to send an intent to another activity
-        onView(withId(R.id.recentTransactionActivityButton)).perform(click());
+        onView(ViewMatchers.withId(R.id.recentTransactionActivityButton)).perform(click());
 
         // Verifies that the DisplayMessageActivity received an intent
         // with the correct package name and message.
